+++ cmd/messaging-server/messaging-server.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/skycoin/skywire/cmd/messaging-server/commands"
-
-func main() {
-	commands.Execute()
-}