+++ cmd/skywire-cli/commands/root.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/skycoin/skywire/cmd/skywire-cli/commands/mdisc"
+	"github.com/skycoin/skywire/cmd/skywire-cli/commands/disc"
 	"github.com/skycoin/skywire/cmd/skywire-cli/commands/node"
 	"github.com/skycoin/skywire/cmd/skywire-cli/commands/rtfind"
 	"github.com/skycoin/skywire/cmd/skywire-cli/commands/tpdisc"
@@ -19,7 +19,7 @@ var rootCmd = &cobra.Command{
 func init() {
 	rootCmd.AddCommand(
 		node.RootCmd,
-		mdisc.RootCmd,
+		disc.RootCmd,
 		rtfind.RootCmd,
 		tpdisc.RootCmd,
 	)