+++ integration/messaging/env-vars.sh
@@ -10,7 +10,7 @@ export RPC_C=$(jq -r ".interfaces.rpc" ./integration/messaging/nodeC.json)
 export CHAT_A=http://localhost:8000/message
 export CHAT_C=http://localhost$(jq -r '.apps [] |select(.app=="skychat")| .args[1] ' ./integration/messaging/nodeC.json)/message
 
-export MSGD=https://messaging.discovery.skywire.skycoin.net
+export DMSGD=https://dmsg.discovery.skywire.skycoin.net
 export TRD=https://transport.discovery.skywire.skycoin.net
 export RF=https://routefinder.skywire.skycoin.net