+++ pkg/setup/node.go
@@ -7,16 +7,14 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/skycoin/skywire/pkg/snet"
-
 	"github.com/skycoin/dmsg"
-
 	"github.com/skycoin/dmsg/cipher"
 	"github.com/skycoin/dmsg/disc"
 	"github.com/skycoin/skycoin/src/util/logging"
 
 	"github.com/skycoin/skywire/pkg/metrics"
 	"github.com/skycoin/skywire/pkg/routing"
+	"github.com/skycoin/skywire/pkg/snet"
 )
 
 // Hop is a wrapper around transport hop to add functionality
@@ -25,7 +23,7 @@ type Hop struct {
 	routeID routing.RouteID
 }
 
-// Node performs routes setup operations over messaging channel.
+// Node performs routes setup operations over dmsg channel.
 type Node struct {
 	Logger   *logging.Logger
 	dmsgC    *dmsg.Client